+++ backend/src/Command/CsvImportCommand.php
@@ -2,15 +2,7 @@
 
 namespace App\Command;
 
-use App\Entity\Offer;
-use App\Entity\Product;
-use App\Repository\OfferRepository;
-use App\Repository\ProductRepository;
-use App\Service\OfferDataHandler;
-use App\Service\ProductDataHandler;
-use App\Service\RecordFinderService;
-use Doctrine\ORM\EntityManagerInterface;
-use League\Csv\Reader;
+use App\Service\CsvReaderService;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -25,27 +17,18 @@ class CsvImportCommand extends Command
     /**
      * @var EntityManagerInterface
      */
-    public $em;
-    public $repository;
-    public $offerDataHandler;
-    public $productDataHandler;
+    public $csvReaderService;
 
     /**
      * CsvImportCommand constructor.
      *
-     * @param EntityManagerInterface $em
-     * @param ProductRepository $repository
-     * @param OfferDataHandler $offerDataHandler
-     * @param ProductDataHandler $productDataHandler
+     * @param CsvReaderService $csvReaderService
      *
      * @throws \Symfony\Component\Console\Exception\LogicException
      */
-    public function __construct(EntityManagerInterface $em, ProductRepository $repository, OfferDataHandler $offerDataHandler, ProductDataHandler $productDataHandler)
+    public function __construct()
     {
-        $this->em = $em;
-        $this->repository = $repository;
-        $this->offerDataHandler = $offerDataHandler;
-        $this->productDataHandler = $productDataHandler;
+        $this->csvReaderService = $csvReaderService;
 
         parent::__construct();
     }
@@ -70,21 +53,9 @@ class CsvImportCommand extends Command
      */
     public function execute(InputInterface $input, OutputInterface $output): int
     {
-        $productUploader = $this->productDataHandler;
-        $offerUploader = $this->offerDataHandler;
+        $csvCommand = $this->csvReaderService;
+        $csvCommand->useData();
 
-        $productReader = Reader::createFromPath(__DIR__ . "/../Data/PRODUCT_DATA.csv", 'r');
-        $productUploader->uploadProductData($productReader);
-
-        $offerReader = Reader::createFromPath(__DIR__ . "/../Data/OFFER_DATA.csv", 'r');
-        $offerUploader->uploadOfferData($offerReader);
-
-        $products = $this->em->getRepository(Product::class)->findAll();
-        foreach ($products as $product) {
-            $searchForGtin = $product->getGtin();
-            $offersFound = $this->em->getRepository(Offer::class)->findOneBy(array('gtin' => $searchForGtin));
-            $product->addOffer($offersFound);
-        }
-        return 0;
+        return Command::SUCCESS;
     }
 }
