+++ frontend/src/services/services/submitOrder.js
@@ -1,6 +1,6 @@
 import { deleteLocally } from "../../lib/localStorage";
 
-export default function handleSubmit(event, orderData, history) {
+export default function submitOrder(event, orderData, history) {
   event.preventDefault();
   const myHeaders = new Headers();
   myHeaders.append("Content-Type", "application/json");
@@ -15,7 +15,6 @@ export default function handleSubmit(event, orderData, history) {
 
   fetch("http://whiskycheck.local/create-order", requestOptions)
     .then((response) => response.text())
-    .then((result) => console.log(result))
     .catch((error) => console.log("error", error));
 
   history.push("/confirmation");
